+++ recipes/cluster.rb
@@ -4,7 +4,24 @@
 #
 
 node.default['percona']['repositories'] = %w(pxc-80)
+node.default['percona']['client']['packages'] =
+  case node['platform_family']
+  when 'debian'
+    if node['percona']['version'].to_f >= 8.0
+      %w(percona-xtradb-cluster-client)
+    else
+      %W(percona-xtradb-cluster-client-#{node['percona']['version']})
+    end
+  when 'rhel'
+    if node['percona']['version'].to_f >= 8.0
+      %w(percona-xtradb-cluster-client)
+    else
+      %W(Percona-XtraDB-Cluster-client-#{node['percona']['version'].tr('.', '')})
+    end
+  end
+
 include_recipe 'percona::package_repo'
+include_recipe 'percona::client'
 
 # Determine and set wsrep_sst_receive_address
 if node['percona']['cluster']['wsrep_sst_receive_interface']