+++ spec/cluster_spec.rb
@@ -11,6 +11,7 @@ describe 'percona::cluster' do
 
   %w(
     percona::package_repo
+    percona::client
     percona::configure_server
     percona::access_grants
   ).each do |r|
@@ -30,6 +31,10 @@ describe 'percona::cluster' do
       it do
         expect(chef_run).to install_package 'percona-xtradb-cluster-56'
       end
+
+      it do
+        expect(chef_run).to install_package 'percona-xtradb-cluster-client-5.6'
+      end
     end
 
     describe 'CentOS' do
@@ -42,6 +47,10 @@ describe 'percona::cluster' do
       it do
         expect(chef_run).to install_package 'Percona-XtraDB-Cluster-56'
       end
+
+      it do
+        expect(chef_run).to install_package 'Percona-XtraDB-Cluster-client-56'
+      end
     end
   end
   describe 'version 5.7' do
@@ -50,6 +59,10 @@ describe 'percona::cluster' do
     describe 'Ubuntu' do
       it do
         expect(chef_run).to install_package 'percona-xtradb-cluster-57'
+      end
+
+      it do
+        expect(chef_run).to install_package 'percona-xtradb-cluster-client-5.7'
       end
     end
 
@@ -63,6 +76,10 @@ describe 'percona::cluster' do
       it do
         expect(chef_run).to install_package 'Percona-XtraDB-Cluster-57'
       end
+
+      it do
+        expect(chef_run).to install_package 'Percona-XtraDB-Cluster-client-57'
+      end
     end
   end
   describe 'version 8.0' do
@@ -71,6 +88,10 @@ describe 'percona::cluster' do
     describe 'Ubuntu' do
       it do
         expect(chef_run).to install_package 'percona-xtradb-cluster-server'
+      end
+
+      it do
+        expect(chef_run).to install_package 'percona-xtradb-cluster-client'
       end
     end
 
@@ -83,6 +104,10 @@ describe 'percona::cluster' do
 
       it do
         expect(chef_run).to install_package 'percona-xtradb-cluster-server'
+      end
+
+      it do
+        expect(chef_run).to install_package 'percona-xtradb-cluster-client'
       end
     end
   end