+++ CHANGELOG.md
@@ -32,6 +32,7 @@ This file is used to list changes made in each version of the percona cookbook.
 - Convert to InSpec tests and refactor test cookbook recipes
 - Don't install abi_version packages on Debian/Ubuntu
 - Standardise files with files in sous-chefs/repo-management
+- Move client package installation for cluster to cluster recipe
 
 ### Removed